+++ src/Foundation/Models/Order.php
@@ -17,7 +17,6 @@ namespace Vanilo\Foundation\Models;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use Konekt\Customer\Models\CustomerProxy;
 use Vanilo\Adjustments\Contracts\Adjustable;
 use Vanilo\Adjustments\Support\HasAdjustmentsViaRelation;
@@ -25,12 +24,12 @@ use Vanilo\Adjustments\Support\RecalculatesAdjustments;
 use Vanilo\Channel\Contracts\Channel;
 use Vanilo\Channel\Models\ChannelProxy;
 use Vanilo\Contracts\Payable;
+use Vanilo\Foundation\Traits\CanBeShipped;
 use Vanilo\Order\Models\Order as BaseOrder;
 use Vanilo\Payment\Contracts\Payment;
 use Vanilo\Payment\Models\PaymentProxy;
 use Vanilo\Shipment\Contracts\Shipment as ShipmentContract;
 use Vanilo\Shipment\Contracts\ShippingMethod;
-use Vanilo\Shipment\Models\ShipmentProxy;
 use Vanilo\Shipment\Models\ShippingMethodProxy;
 
 /**
@@ -47,6 +46,7 @@ use Vanilo\Shipment\Models\ShippingMethodProxy;
  */
 class Order extends BaseOrder implements Payable, Adjustable
 {
+    use CanBeShipped;
     use HasAdjustmentsViaRelation;
     use RecalculatesAdjustments;
 
@@ -153,24 +153,5 @@ class Order extends BaseOrder implements Payable, Adjustable
     public function payments()
     {
         return $this->morphMany(PaymentProxy::modelClass(), 'payable');
-    }
-
-    public function shipments(): MorphToMany
-    {
-        return $this->morphToMany(ShipmentProxy::modelClass(), 'shippable');
-    }
-
-    public function addShipment(ShipmentContract $shipment): static
-    {
-        $this->shipments()->save($shipment);
-
-        return $this;
-    }
-
-    public function addShipments(ShipmentContract ...$shipments)
-    {
-        $this->shipments()->saveMany($shipments);
-
-        return $this;
     }
 }