+++ Changelog.md
@@ -14,6 +14,7 @@
 - Added the extended `CartItem` and `OrderItem` models to Foundation that are `Adjustable`
 - Added the `fromKeyValuePairs` static factory and `getDetail()` methods to the DetailedAmount Dto class
 - Added the `carrier_cost`, `label_url` and `label_base64` fields to the shipments table/model
+- Added feature to assign shipments to order items, and vice versa (Foundation)
 
 ## 3.6.2
 ##### 2023-03-12